refactor(mcp): inline the typed-field helper into patch_entity, its only caller
